requirements.txt
pytorch-lightning >= 1.0.0rc2
torch >= 1.3.0
torchvision >= 0.6.0

# yolo api
ultralytics

# config hydra 
hydra-core

# for cross validation
scikit-learn